build.gradle 주석 제거, detachView 메소드 제거, tag제거, onDestroy 메소드 Override안함
