+++ app/build.gradle
@@ -26,8 +26,8 @@ dependencies {
     })
     compile 'com.android.support:appcompat-v7:25.3.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.jakewharton:butterknife:8.7.0' //ButterKnife
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.7.0'    //ButterKnife
-    compile 'com.android.volley:volley:1.0.0'   //volley Library
+    compile 'com.jakewharton:butterknife:8.7.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:8.7.0'
+    compile 'com.android.volley:volley:1.0.0'
     testCompile 'junit:junit:4.12'
 }
