+++ app/src/main/java/com/hpcnt/imageloader/main/MainActivity.java
@@ -1,0 +1,33 @@
+package com.hpcnt.imageloader.main;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.widget.TextView;
+
+import com.android.volley.toolbox.Volley;
+import com.hpcnt.imageloader.R;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
+public class MainActivity extends AppCompatActivity implements MainContract.View {
+    private MainPresenter mainPresenter;
+    @BindView(R.id.textView)
+    TextView textView;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        ButterKnife.bind(this);
+        mainPresenter = new MainPresenter();
+        mainPresenter.attachView(this);
+        mainPresenter.setVolleyQueue(Volley.newRequestQueue(getApplicationContext()));
+        mainPresenter.getHtml("http://www.naver.com");
+    }
+
+    @Override
+    public void setTextView(String text) {
+        textView.setText(text);
+    }
+}
